# BlockMerge optimisation