--- a/casper/src/main/scala/coop/rchain/casper/blocks/merger/package.scala
+++ b/casper/src/main/scala/coop/rchain/casper/blocks/merger/package.scala
@@ -1,8 +1,0 @@
-package coop.rchain.casper.blocks
-
-import coop.rchain.metrics.Metrics
-
-package object merger {
-  val MergingMetricsSource: Metrics.Source =
-    Metrics.Source(Metrics.BaseSource, "merging")
-}