This PR continues work on block merge.

These are results for merging 19 blocks from peers. This is very first merge, looks like JVM were not warmed up, so numbers are slightly bigger than with averaging.

```2021-02-25 17:33:42,307 [INFO ] [node-runner-85      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 278 consumes [1.051922584 sec]
2021-02-25 17:33:42,318 [INFO ] [node-runner-85      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 310 consumes [1.148264846 sec]
2021-02-25 17:33:42,369 [INFO ] [node-runner-64      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 350 consumes [1.137562705 sec]
2021-02-25 17:33:42,391 [INFO ] [node-runner-64      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 302 consumes [1.212630647 sec]
2021-02-25 17:33:42,426 [INFO ] [node-runner-82      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 306 consumes [1.193673457 sec]
2021-02-25 17:33:42,474 [INFO ] [node-runner-50      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 314 consumes [1.225670189 sec]
2021-02-25 17:33:42,477 [INFO ] [node-runner-81      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 338 consumes [1.266100753 sec]
2021-02-25 17:33:42,512 [INFO ] [node-runner-81      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 286 consumes [1.256947788 sec]
2021-02-25 17:33:42,515 [INFO ] [node-runner-82      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 354 consumes [1.317729183 sec]
2021-02-25 17:33:42,537 [INFO ] [node-runner-50      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 298 consumes [1.26495599 sec]
2021-02-25 17:33:42,545 [INFO ] [node-runner-85      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 342 consumes [1.326230389 sec]
2021-02-25 17:33:42,547 [INFO ] [node-runner-91      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 334 consumes [1.318260285 sec]
2021-02-25 17:33:42,556 [INFO ] [node-runner-59      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 322 consumes [1.296293919 sec]
2021-02-25 17:33:42,568 [INFO ] [node-runner-85      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 282 consumes [1.422848821 sec]
2021-02-25 17:33:42,568 [INFO ] [node-runner-90      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 346 consumes [1.291053695 sec]
2021-02-25 17:33:42,575 [INFO ] [node-runner-91      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 330 consumes [1.30369556 sec]
2021-02-25 17:33:42,582 [INFO ] [node-runner-87      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 318 consumes [1.205777585 sec]
2021-02-25 17:33:42,584 [INFO ] [node-runner-58      ] [c.r.r.m.i.DiffStateMerger    ]- compute changes: 32 produces + 326 consumes [1.198131412 sec]
2021-02-25 17:33:47,752 [INFO ] [node-runner-53      ] [c.r.r.m.i.DiffStateMerger    ]- process trieActions [5.150237549 sec]
2021-02-25 17:33:47,752 [INFO ] [node-runner-53      ] [c.r.c.b.m.CasperDagMerger$   ]- Merging done. [8.342220589 sec]
```

NOTE: process trieActions takes quite a lot of time - this is because 
https://github.com/rchain/rchain/blob/5a9b6d418ebb4cc86781ae69070d629491a6414a/rspace/src/main/scala/coop/rchain/rspace/history/HistoryInstances.scala#L318-L327
here history actions processing is sequential. It makes sense at least for testing environment to use parallel processing.

UPD. with parallel processing of trieActions `process trieActions [1.443238755 sec]`

For historical purposes, let me record here results of testing on network of 10, 20, 30 and 40 nodes at this commit https://github.com/rchain/rchain/pull/3312/commits/1748cc6d7ea98d1ad75f80bf891e3d685b8be9b5, **but with parallel processing of trie actions**. This is quite raw estimate, but this is what we have now. 
<p align="center"><img width="591" alt="Screenshot 2021-02-26 at 18 27 45" src="https://user-images.githubusercontent.com/1746367/109319803-5567ed80-7860-11eb-9fb6-cccacb05b495.png"></p>

Looks like conflict detection part grows exponentially, and thats reasonable - now we check all branches combinations . One solution is to use map-reduce, also caching might be implemented.

closes https://github.com/rchain/rchain/issues/3308
closes https://github.com/rchain/rchain/issues/3277 (except reimplement tests disabled in 53578aa, if applicable)